+++ server/src/main/java/com/ticketer/dtos/ItemDto.java
@@ -3,9 +3,9 @@ package com.ticketer.dtos;
 import java.util.Map;
 
 public record ItemDto(
-        String name,
-        int price,
-        boolean available,
-        Map<String, SideDto> sides,
-        java.util.List<String> sideOrder) {
+                String name,
+                long price,
+                boolean available,
+                Map<String, SideDto> sides,
+                java.util.List<String> sideOrder) {
 }
