+++ server/src/main/java/com/ticketer/dtos/SideDto.java
@@ -1,6 +1,6 @@
 package com.ticketer.dtos;
 
 public record SideDto(
-        int price,
-        boolean available) {
+                long price,
+                boolean available) {
 }
